Thanks for your contribution and we appreciate it a lot. 

## 1. Motivation

Add optimized kernels for ms_deform_attn_forwrad/backwrad ops.

## 2. Modification

bangc-ops/kernels/ms_deform_attn_backward/ms_deform_attn_backward.cpp
bangc-ops/kernels/ms_deform_attn_backward/ms_deform_attn_backward.h
bangc-ops/kernels/ms_deform_attn_backward/ms_deform_attn_backward_fast_union1.mlu
bangc-ops/kernels/ms_deform_attn_forward/ms_deform_attn_forward.mlu
bangc-ops/kernels/ms_deform_attn_forward/ms_deform_attn_utils.h
bangc-ops/kernels/ms_deform_attn_forward/msda_forward_fast_union1.mlu

## 3. Test Report

If you want to know how to do operator testing, you can see [GTest-User-Guide-zh](https://github.com/Cambricon/mlu-ops/blob/master/docs/GTest-User-Guide-zh.md).

### 3.1 Modification Details

#### 3.1.1 Accuracy Acceptance Standard

For static threshold standard details, see: [MLU-OPS Accuracy Acceptance Standard](https://github.com/Cambricon/mlu-ops/blob/master/docs/MLU-OPS-Accuracy-Acceptance-Standard.md).

- [ ] diff1: diff1 <= 3e-3
- [ ] diff2: diff2 <= 3e-3

#### 3.1.2 Operator Scheme checklist

|     No.        |                 Details              |            Check Results             |
|----------------|--------------------------------------|--------------------------------------|
|        1       |Supported hardware                    |             MLU370<br>MLU590         |
|        2       |Job types                             |           U1       |
|        3       |Layouts                               |          ARRAY       |
|        4       |Whether multi-dimensions are supported|                 No                     |
|        5       |Whether element zero is supported     |                    Yes                  |
|        6       |Data type(half/float)                 |            float           |
|        7       |Whether there is size limit           |                 No                     |

#### 3.1.3 New Feature Test

Since this ops is merged in other repo, we only focus on whether this ops works well.

For all 3 cases of network, this ops passed unit test.

### 3.2 Summary Analysis

For Accurary

Test Point | Operation | Quantity | Comment
-- | -- | -- | --
 | accurary| ms_deform_attn_backward | 通过 | [==========] 123 test cases from 1 test suite ran. (195191 ms total)<br />[  PASSED  ] 123 test cases.
 | nan/inf| ms_deform_attn_backward | 通过 | [ SUMMARY  ] Total 33 cases of 1 op(s).<br />ALL PASSED.<br />[==========] 33 test cases from 1 test suite ran. (20231 ms total)<br />[  PASSED  ] 33 test cases.
 | accurary | ms_deform_attn_forward | 通过 | [ SUMMARY  ] Total 31 cases of 1 op(s).<br />ALL PASSED.<br />[==========] 31 test cases from 1 test suite ran. (39895 ms total)<br />[  PASSED  ] 31 test cases.

For Efficiency

The new added fast kernels is only for MLU590. The performance of typical cases on MLU590-H8:
|Operation|Mlu_hardware_time baseline(us)|Mlu_hardware_time this (us)|Mlu_interface_time(us)|Mlu_io_efficiency|Mlu_compute_efficiency|Data_type|Shape|
|-------|----|----|----|-----|----|----|----|
|  ms_deform_attn_backward | 132000 |  71113 |  51  |   0.5% | 14%   |  float  | data_value: [6, 30825, 8, 32]<br />data_spatial_shapes: [4, 2]<br />data_level_start_index: [4]<br />data_sampling_loc: [6, 9664, 8, 4, 8, 2]<br />data_attn_weight: [6, 9664, 8, 4, 8]<br />grad_output: [6, 9664, 8, 32]    |
|  ms_deform_attn_backward   | 25078 |  14953  |   18.9 |   1.0% | 11.4%   | float  |  data_value: [2, 40000, 8, 32]data_spatial_shapes: [1, 2]<br />data_level_start_index: [1]<br />data_sampling_loc: [2, 40000, 8, 1, 4, 2]<br />data_attn_weight: [2, 40000, 8, 1, 4]<br />grad_output: [2, 40000, 8, 32]<br />grad_value: [2, 40000, 8, 32]   |
|  ms_deform_attn_forward  | 28000 | 9411 |  51  |   40.8% | 0.7%   |  float  | data_value: [6, 30825, 8, 32]<br />data_spatial_shapes: [4, 2]<br />data_level_start_index: [4]<br />data_sampling_loc: [6, 9664, 8, 4, 8, 2]<br />data_attn_weight: [6, 9664, 8, 4, 8]<br />value_output: [6, 9664, 8, 32]    |
|  ms_deform_attn_forward   | 16500 | 2248  |   18.9 |   31.0% | 0.5%   | float  |  data_value: [2, 40000, 8, 32]data_spatial_shapes: [1, 2]<br />data_level_start_index: [1]<br />data_sampling_loc: [2, 40000, 8, 1, 4, 2]<br />data_attn_weight: [2, 40000, 8, 1, 4]<br />grad_output: [2, 40000, 8, 32]<br />output_value: [2, 40000, 8, 32]   |